Separate entity header from sighting core
