+++ src/pages/sighting/SightingCore.jsx
@@ -1,55 +1,27 @@
 import React, { useMemo, useState } from 'react';
 import { useHistory } from 'react-router-dom';
-import { useIntl } from 'react-intl';
-import { get, map } from 'lodash-es';
+import { get } from 'lodash-es';
 import { useQueryClient } from 'react-query';
 
 import {
   getAGSQueryKey,
   getSightingQueryKey,
 } from '../../constants/queryKeys';
-import timePrecisionMap from '../../constants/timePrecisionMap';
-import defaultSightingSrc from '../../assets/defaultSighting.png';
 import MainColumn from '../../components/MainColumn';
-import Link from '../../components/Link';
-import Tabs from '../../components/Tabs';
-import Text from '../../components/Text';
 import LoadingScreen from '../../components/LoadingScreen';
 import SadScreen from '../../components/SadScreen';
-import MoreMenu from '../../components/MoreMenu';
 import ConfirmDelete from '../../components/ConfirmDelete';
-import EntityHeader from '../../components/EntityHeader';
 import useDocumentTitle from '../../hooks/useDocumentTitle';
 import useDeleteSighting from '../../models/sighting/useDeleteSighting';
 import useDeleteAssetGroupSighting from '../../models/assetGroupSighting/useDeleteAssetGroupSighting';
 import useSightingFieldSchemas from '../../models/sighting/useSightingFieldSchemas';
-import { formatDateCustom } from '../../utils/formatters';
+import SightingEntityHeader from './SightingEntityHeader';
 import Annotations from './Annotations';
 import Photographs from './Photographs';
 import OverviewContent from './OverviewContent';
 import SightingHistoryDialog from './SightingHistoryDialog';
 import CommitBanner from './CommitBanner';
-import FeaturedPhoto from '../../components/FeaturedPhoto';
 import Encounters from './encounters/Encounters';
-
-const tabItems = [
-  {
-    labelId: 'OVERVIEW',
-    value: 'overview',
-  },
-  {
-    labelId: 'PHOTOGRAPHS',
-    value: 'photographs',
-  },
-  {
-    labelId: 'ANNOTATIONS',
-    value: 'annotations',
-  },
-  {
-    labelId: 'ANIMALS',
-    value: 'individuals',
-  },
-];
 
 export default function SightingCore({
   data,
@@ -60,7 +32,6 @@ export default function SightingCore({
   id,
 }) {
   const history = useHistory();
-  const intl = useIntl();
   const queryClient = useQueryClient();
 
   const fieldSchemas = useSightingFieldSchemas();
@@ -100,24 +71,6 @@ export default function SightingCore({
     [data, fieldSchemas],
   );
 
-  const dataForFeaturedPhoto = useMemo(
-    () => {
-      const assets = get(data, ['assets'], []);
-      const modifiedAssets = map(
-        assets,
-        asset => {
-          return { ...asset, altText: asset?.filename };
-        },
-        [],
-      );
-      return {
-        ...data,
-        assets: modifiedAssets,
-      };
-    },
-    [data],
-  );
-
   useDocumentTitle(`Sighting ${id}`, { translateMessage: false });
 
   const [historyOpen, setHistoryOpen] = useState(false);
@@ -143,24 +96,6 @@ export default function SightingCore({
   if (error) return <SadScreen variant="genericError" />;
 
   const assets = get(data, 'assets', []);
-
-  const sightingTime = data?.time;
-  const sightingTimeSpecificity = data?.timeSpecificity;
-  const formatSpecification = get(
-    timePrecisionMap,
-    [sightingTimeSpecificity, 'prettyFormat'],
-    'yyyy-MM-dd',
-  );
-  const sightingDisplayDate = formatDateCustom(
-    sightingTime,
-    formatSpecification,
-  );
-
-  const sightingCreator = data?.creator;
-  const creatorName =
-    sightingCreator?.full_name ||
-    intl.formatMessage({ id: 'UNNAMED_USER' });
-  const creatorUrl = `/users/${sightingCreator?.guid}`;
 
   return (
     <MainColumn fullWidth>
@@ -199,57 +134,15 @@ export default function SightingCore({
         }
         messageId="CONFIRM_DELETE_SIGHTING_DESCRIPTION"
       />
-      <EntityHeader
-        renderAvatar={
-          <FeaturedPhoto
-            data={pending ? null : dataForFeaturedPhoto}
-            loading={loading}
-            refreshData={refreshData}
-            defaultPhotoSrc={defaultSightingSrc}
-            sightingId={data?.guid}
-          />
-        }
-        name={intl.formatMessage(
-          { id: 'ENTITY_HEADER_SIGHTING_DATE' },
-          { date: sightingDisplayDate },
-        )}
-        renderTabs={
-          <Tabs
-            value={activeTab.replace('#', '')}
-            onChange={(_, newValue) => {
-              window.location.hash = newValue;
-            }}
-            items={tabItems}
-          />
-        }
-        renderOptions={
-          <div style={{ display: 'flex' }}>
-            {/* <Button id="SUBSCRIBE" display="primary" /> */}
-            <MoreMenu
-              menuId="sighting-actions"
-              items={[
-                {
-                  id: 'view-history',
-                  onClick: () => setHistoryOpen(true),
-                  label: 'View history',
-                },
-                {
-                  id: 'delete-sighting',
-                  onClick: () => setDeleteDialogOpen(true),
-                  label: 'Delete sighting',
-                },
-              ]}
-            />
-          </div>
-        }
-      >
-        {sightingCreator && (
-          <Text variant="body2">
-            {intl.formatMessage({ id: 'REPORTED_BY' })}
-            <Link to={creatorUrl}>{creatorName}</Link>
-          </Text>
-        )}
-      </EntityHeader>
+      <SightingEntityHeader
+        activeTab={activeTab}
+        data={data}
+        loading={loading}
+        pending={pending}
+        id={id}
+        setHistoryOpen={setHistoryOpen}
+        setDeleteDialogOpen={setDeleteDialogOpen}
+      />
       <CommitBanner
         sightingId={id}
         pending={pending}
